Make box-shadow in cards permanent
We were using it on hover because the cards were links. However,
displaying it all the time meant cards were overlapping.

So we have to add a gap between cards, but then we also have to
recalculate the width so the gap is taken into account. Since the gap
depends on the `card.columns` attribute, we're including it in a HTML
data attribute.